feat: seperate node and browser loggers
