+++ packages/common/src/index.ts
@@ -8,6 +8,7 @@ export * from './logger-provider'
 export * from './logger-provider-old' // TODO remove this
 export * from './loggers'
 export * from './networks'
+export * from './logger-base'
 export * from './pinning'
 export * from './unreachable-case-error'
 export * from './running-state-like'
